requirements.txt
gymnasium
numpy
torch  # Adjust this version depending on your device
